_classification_support.py -  Refactored checks into ClassificationSupport, precision/recall calculation into PrecisionRecallSupport. precision.py and recall.py - use PrecisionRecallSupport
